[refactor] #13: 신고 타입 type/report 패키지로 이동
